requirements.txt
apache-airflow==2.3.0
apache-airflow-providers-docker==2.3.0
apache-airflow-providers-postgres==2.3.0
psycopg2-binary==2.9.9